% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\docType{data}
\name{eigenvalues}
\alias{eigenvalues}
\title{eigenvalues}
\format{A data frame with 20 rows and 1 variable:
\describe{
  \item{eigen_values}{calculated eigennvalues}
}}
\source{
{Machiela Lab}
}
\usage{
eigenvalues
}
\description{
A sample data set containing eivenvalues
}
\examples{
  eigenvalues
\dontrun{genome_values <- eigenvalues
         values <- c(genome_values)$eigen_values
}

}
\keyword{datasets}
