% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues_1000G.R
\docType{data}
\name{eigenvalues_1000G}
\alias{eigenvalues_1000G}
\title{eigenvalues_1000G}
\format{A data frame with 20 rows and 1 variable:
\describe{
  \item{eigen_values}{calculated eigennvalues}
}}
\source{
{Machiela Lab}
}
\usage{
eigenvalues_1000G
}
\description{
A sample data set containing eivenvalues
}
\examples{
  eigenvalues_1000G
\donttest{genome_values <- eigenvalues
          values <- c(genome_values)$eigen_values
}

}
\keyword{datasets}
