% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matches.R
\name{plot_matches}
\alias{plot_matches}
\title{Function to plot matches}
\usage{
plot_matches(
  data = NULL,
  x_var = NULL,
  y_var = NULL,
  case_control = NULL,
  line = T
)
}
\arguments{
\item{data}{PCAmatchR output}

\item{x_var}{Principal component 1}

\item{y_var}{Principal component 2}

\item{case_control}{Case or control status}

\item{line}{line}
}
\value{
A plot
}
\description{
Function to plot matches
}
\examples{
\dontrun{plot_matches(data=PCAmatchR_output,
                     x_var="PC1",
                     y_var="PC2",
                     case_control="case",
                     line=T)}

}
