% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues_1000G.R
\docType{data}
\name{eigenvalues_1000G}
\alias{eigenvalues_1000G}
\title{Eigenvalues of 2504 individuals from the 1000 Genome Project}
\format{A data frame with 20 rows and 1 variable:
\describe{
  \item{eigen_values}{calculated eigennvalues}
}}
\source{
{Machiela Lab}
}
\usage{
eigenvalues_1000G
}
\description{
A sample dataset containing the first 20 eigenvalues
calculated from 2504 individuals in the Phase 3 data
release of the 1000 Genomes Project. The principal
component analysis was conducted using PLINK.
}
\examples{
  eigenvalues_1000G
\donttest{genome_values <- eigenvalues
          values <- c(genome_values)$eigen_values
}

}
\keyword{datasets}
