% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maker.R
\name{plot_maker}
\alias{plot_maker}
\title{Function to plot matches from match_maker output}
\usage{
plot_maker(
  data = NULL,
  x_var = NULL,
  y_var = NULL,
  case_control = NULL,
  line = T
)
}
\arguments{
\item{data}{match_maker output}

\item{x_var}{Principal component 1}

\item{y_var}{Principal component 2}

\item{case_control}{Case or control status}

\item{line}{draw line}
}
\value{
A plot
}
\description{
Function to plot matches from match_maker output
}
\examples{
\dontrun{plot_maker(data=match_maker_output,
                     x_var="PC1",
                     y_var="PC2",
                     case_control="case",
                     line=T)}

}
